+++ app/src/main/java/com/tyron/code/ui/wizard/WizardFragment.java
@@ -107,9 +107,9 @@ public class WizardFragment extends Fragment {
 
     setEnterTransition(new MaterialSharedAxis(MaterialSharedAxis.X, false));
     setExitTransition(new MaterialSharedAxis(MaterialSharedAxis.X, true));
-      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-          mUseInternalStorage = Environment.isExternalStorageManager();
-      }
+//      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+//          mUseInternalStorage = Environment.isExternalStorageManager();
+//      }
 
       mPermissionLauncher =
         registerForActivityResult(
@@ -173,7 +173,7 @@ public class WizardFragment extends Fragment {
 
     mAdapter = new WizardTemplateAdapter();
     mRecyclerView.setAdapter(mAdapter);
-    PreferenceManager.getDefaultSharedPreferences(requireContext()).edit().putString(SharedPreferenceKeys.PROJECT_SAVE_PATH,"/storage/emulated/0/CodeAssistProjects").apply();
+//    PreferenceManager.getDefaultSharedPreferences(requireContext()).edit().putString(SharedPreferenceKeys.PROJECT_SAVE_PATH,"/storage/emulated/0/CodeAssistProjects").apply();
 
     initDetailsView();