+++ changelogs/0.2.0-ALPHA02/changelog.md
@@ -4,6 +4,8 @@
 - added ` Java 17 ` completion and compilation (full support)
 - added ` ColorPicker ` for Icon Manager
 - added some attribute snippets in `xml-completion`
+- integrated new Kotlin-lsp based on CodeOnTheGo Kotlin-lsp supports ( JetPack Compose )
+- integrated Compose Preview ( unstable )
 
 ### fixes
 - skipping compile kotlin files